- [BE] feat(world): 뉴스/이벤트 스키마 생성 및 공통 에러 코드 추가 (S14P21C103-191)

See merge request s14-fintech-finance-sub1/S14P21C103!63

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/world/housing/GameContractReview.java
@@ -28,7 +28,7 @@ public class GameContractReview {
     private Long id;
 
     @Column(name = "game_session_id", nullable = false)
-    private Long gameSessionId;
+    private Integer gameSessionId;
 
     @Column(name = "property_id", nullable = false)
     private Long propertyId;
@@ -53,7 +53,7 @@ public class GameContractReview {
     private LocalDateTime reviewedAt;
 
     private GameContractReview(
-        Long gameSessionId,
+        Integer gameSessionId,
         Long propertyId,
         ContractReviewStatus reviewStatus,
         List<String> checkedTraps,
@@ -71,7 +71,7 @@ public class GameContractReview {
     }
 
     public static GameContractReview create(
-        Long gameSessionId,
+        Integer gameSessionId,
         Long propertyId,
         ContractReviewStatus reviewStatus,
         List<String> checkedTraps,
