+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/world/housing/GameHousingRepository.java
@@ -3,7 +3,7 @@ package io.ssafy.p.j14c103.homerun.domain.world.housing;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameHousingRepository extends JpaRepository<GameHousing, Long> {
+public interface GameHousingRepository extends JpaRepository<GameHousing, Integer> {
 
-    Optional<GameHousing> findByGameSessionId(Long gameSessionId);
+    Optional<GameHousing> findByGameSessionId(Integer gameSessionId);
 }
